+++ src/assets/icons/spending/acmdation_clicked.png
