+++ src/assets/icons/spending/etc_clicked.png
