+++ src/assets/icons/spending/food.png
