+++ src/assets/icons/spending/food_clicked.png
