+++ src/assets/icons/spending/shopping_clicked.png
