+++ src/assets/icons/spending/tour_clicked.png
