+++ src/assets/icons/spending/transport_clicked.png
