+++ src/components/OptionButton.js
@@ -39,38 +39,38 @@ const CategoryOptionButton = () => {
         {
             id: 1,
             text: "교통",
-            image: require('../assets/images/transport.png'),
-            image_clicked: require('../assets/images/transport_clicked.png'),
+            image: require('../assets/icons/spending/transport.png'),
+            image_clicked: require('../assets/icons/spending/transport_clicked.png'),
         },
         {
             id: 2,
             text: "식사",
-            image: require('../assets/images/food.png'),
-            image_clicked: require('../assets/images/food_clicked.png'),
+            image: require('../assets/icons/spending/food.png'),
+            image_clicked: require('../assets/icons/spending/food_clicked.png'),
         },
         {
             id: 3,
             text: "관광",
-            image: require('../assets/images/tour.png'),
-            image_clicked: require('../assets/images/tour_clicked.png'),
+            image: require('../assets/icons/spending/tour.png'),
+            image_clicked: require('../assets/icons/spending/tour_clicked.png'),
         },
         {
             id: 4,
             text: "쇼핑",
-            image: require('../assets/images/shopping.png'),
-            image_clicked: require('../assets/images/shopping_clicked.png'),
+            image: require('../assets/icons/spending/shopping.png'),
+            image_clicked: require('../assets/icons/spending/shopping_clicked.png'),
         },
         {
             id: 5,
             text: "숙소",
-            image: require('../assets/images/acmdation.png'),
-        image_clicked: require('../assets/images/acmdation_clicked.png'),
+            image: require('../assets/icons/spending/acmdation.png'),
+        image_clicked: require('../assets/icons/spending/acmdation_clicked.png'),
     },
         {
             id: 6,
             text: "기타",
-            image: require('../assets/images/etc.png'),
-            image_clicked: require('../assets/images/etc_clicked.png'),
+            image: require('../assets/icons/spending/etc.png'),
+            image_clicked: require('../assets/icons/spending/etc_clicked.png'),
         },
     ]
