+++ src/navigations/Stack.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { createStackNavigator } from '@react-navigation/stack';
 import { Image, TouchableOpacity } from 'react-native';
-import BackDark from '../assets/images/back_dark.png';
-import BackBright from '../assets/images/back_bright.png';
-import HeartEmpty from '../assets/images/heart_empty.png';
+import BackDark from '../assets/icons/ai/back_dark.png';
+import BackBright from '../assets/icons/ai/back_bright.png';
+import HeartEmpty from '../assets/icons/ai/heart_empty.png';
 
 // 테스트용 데이터
 // const ScreenList = [
